+++ repentogon/LuaEntityFamiliar.cpp
@@ -25,7 +25,7 @@ int Lua_FamiliarTryAimAtMarkedTarget(lua_State* L)
 {
 	Entity_Familiar* fam = lua::GetUserdata<Entity_Familiar*>(L, 1, lua::Metatables::ENTITY_FAMILIAR, "EntityFamiliar");
 	Vector* aimDirection = lua::GetUserdata<Vector*>(L, 2, lua::Metatables::VECTOR, "Vector");
-	int direction = luaL_checkinteger(L, 3);
+	int direction = (int) luaL_checkinteger(L, 3);
 	Vector* unkVec = NULL;
 	if (lua_gettop(L) == 4 && !lua_isnil(L, 4)) {
 		unkVec = lua::GetUserdata<Vector*>(L, 4, lua::Metatables::VECTOR, "Vector");
